Don't try to support installing over browser-proxy
The diversions seem to make it impossible to support directly installing
over browser-proxy without a dummy package, and that's not worth it
given that we're not in any repositories yet.  So just conflict with
browser-proxy to make sure no one tries to install directly over, and
tell users to remove browser-proxy instead.