Fremantle: clear hildon-control-panel's cached image on remove
On Fremantle, hildon-control-panel creates a screenshot of itself, which
is displayed during startup to create the illusion of a faster startup
time (sigh).  Of course, this cache doesn't appear to be regenerated
when a control panel icon is removed
(https://bugs.maemo.org/show_bug.cgi?id=9512).  Wonderful.

Work around by removing the cached image in the postrm.