Clean up configuration again
Commit ec8b58af... ("Refactor configuration") still has too much
boilerplate code and leaves too many places that have to be modified to
add a new config option.

To fix this, add a new array of config options with name and type of
information stored, and then add an array of pointers to the elements of
struct swb_config with indices matching the list of config options to
struct swb_config.  This consolidates all the work needed to add new
config options to config.h and config.c, and allows us to replace the
boilerplate code with loops over the array of config options.