Introduce command-line config utility
Add a simple command-line configuration utility,
browser-switchboard-config, which can be used to query or set
browser-switchboard config settings.

The primary purpose of this tool is to allow browser vendors to provide
an option in their browsers to set themselves as the default browser,
via something like this (shell-like pseudocode):
	if [ -x /usr/bin/browser-switchboard-config ]; then
		if [ `browser-switchboard-config -b` = "mybrowser" ]; then
			display "MyBrowser is the default browser."
		else
			button "Make MyBrowser the default browser" action "browser-switchboard-config -s -b mybrowser"
		fi
	else
		display "Install Browser Switchboard to make MyBrowser the default browser"
	fi
Browser vendors: setting yourself as the default browser without
prompting the user first (via package maintainer scripts, for example)
is strongly discouraged.