Listen on D-Bus system bus as well as session bus
Discussion on https://bugs.maemo.org/show_bug.cgi?id=8369 (Hermes
doesn't work when Browser Switchboard is installed) revealed that
MicroB listens on the D-Bus system bus as well as the session bus.
While it's not appropriate for applications to request opening a
browser via the system bus (which is supposed to be for requesting
system services, not applications), the fact that MicroB listens
there means that we need to too.

This fixes Browser Switchboard bug 5910.  Thanks to Andrew Flegg and
Marcin Juszkiewicz for providing the information needed to track down
this bug.