Update package description
Emanuele Cassioli pointed out that the package description doesn't
mention that we ship a control panel applet to configure the default
browser, meaning that users who never bother to look for the
documentation may not know where to start.  Fix this.

Also, remove the mention of browserd from the description; the
functionality will remain, but this is no longer one of the main selling
points of the package.