Config UI: Only offer the user browsers that are installed
Check to see which of the known browsers is installed, and only offer
the user choices for the ones that are.  If the currently configured
default browser is not installed (e.g. if the user uninstalled the
browser without changing the default), use the default default browser
like browser-switchboard does.

This fixes Browser Switchboard bug 5469:
https://garage.maemo.org/tracker/index.php?func=detail&aid=5469&group_id=1159&atid=4348