Fremantle: Prestart MicroB if appropriate when reconfiguring browser-switchboard
If, in changing the browser-switchboard configuration, we go from a
configuration where MicroB is not left running to one where it is, we
should start MicroB in the background in order to make sure that the
MicroB browser window comes up quickly when requested.  Make a
best-effort attempt at this in both the command-line utility and the
GUI, refactoring to give a swb_reconfig() function that can be shared
between the two along the way.

Ideally, we'd also kill MicroB when making a config change in the other
direction, but we don't have an easy way of knowing whether MicroB is
actually in use and we don't want to kill MicroB if it's in use at the
time.