Add an Xsession.post script for Fremantle to launch browser-switchboard on session startup
Fremantle MicroB autostarts by default and listens on both the D-Bus
system and session buses, so applications may send requests on either
bus (even though those that send requests on the system bus should be
fixed).  We therefore need to duplicate this behavior, but we cannot
rely on D-Bus to launch us to handle a request coming in on the system
bus, since the system bus runs as the "messagebus" user, and we need to
run with the UID of the logged-in user.  A complete fix for Browser
Switchboard bug 5910 ("browser-switchboard should listen on D-Bus system
bus as well") on Fremantle therefore requires that browser-switchboard
be autostarted when the desktop environment comes up.

We arrange for this the same way browserd does on Fremantle -- with an
Xsession.post script that runs on X session startup.  Also, we prestart
MicroB if the current browser-switchboard configuration is to leave it
running in the background, to make opening the first MicroB window
faster.

Because we may start the MicroB browser process, this script needs to
be inserted after browserd's in the Xsession.post run order.