Update postinst
We can now coexist with a running MicroB browser process, so don't try
to kill off browser in the postinst.  On the other hand, we now need
browser-switchboard running at all times in order for things to work
correctly (especially if there is a MicroB browser process running), so
start browser-switchboard as the user.  We assume the Maemo default
user, since we have no good way of checking this, and on a regular
device (e.g. not the SDK) it's always true.