Fremantle: Use open_new_window instead of load_url when passing URL to MicroB
With the changes to MicroB launching for Fremantle, there are now cases
where we are passing a D-Bus request to a MicroB process with windows
open, instead of just asking for a browser session to be started.  In
the latter case, load_url and open_new_window are equivalent, but in the
former, they are not -- load_url loads the URL in an existing browser
window.

Fully emulating MicroB's behavior here now requires keeping track of
whether the request we got was load_url or open_new_window and
requesting the same of MicroB, but for now, do what most users want and
what we do with other browsers -- always open a new window for the URL.