Provide open_new_window and load_url methods with boolean fullscreen argument
MicroB implements an optional boolean argument to the open_new_window
and load_url methods in the com.nokia.osso_browser interface, which
allows the application to request that the new page be loaded in a
fullscreen window.  We need to provide this as well -- otherwise,
applications that use the argument will receive an error.

Provide versions of open_new_window and load_url with a fullscreen
argument, which is ignored for the moment.  It's not going to be
possible to implement the fullscreen behavior for all supported
browsers, but we should implement it for MicroB, at least.

Fixes the launching of the Flickr new account page from the Sharing
control panel in Fremantle:
https://garage.maemo.org/tracker/index.php?func=detail&aid=6444&group_id=1159&atid=4348
Thanks to Tom G for reporting.